# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MedVision AI",
    "description": "MedVision AI \u2014 an AI-powered healthcare education platform combining image, voice, and text understanding with a Retrieval-Augmented Generation medical knowledge base. AI output is informational only and is not a substitute for a licensed physician.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/diagnosis/image": {
      "post": {
        "tags": [
          "Image Diagnosis"
        ],
        "summary": "Diagnose Image",
        "operationId": "diagnose_image_api_v1_diagnosis_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_diagnose_image_api_v1_diagnosis_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDiagnosisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/diagnosis/image/{image_id}": {
      "get": {
        "tags": [
          "Image Diagnosis"
        ],
        "summary": "Get Image Result",
        "operationId": "get_image_result_api_v1_diagnosis_image__image_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDiagnosisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/image/series/{series_id}/progression": {
      "get": {
        "tags": [
          "Image Diagnosis"
        ],
        "summary": "Get Progression",
        "operationId": "get_progression_api_v1_diagnosis_image_series__series_id__progression_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/voice": {
      "post": {
        "tags": [
          "Voice Diagnosis"
        ],
        "summary": "Diagnose Voice",
        "operationId": "diagnose_voice_api_v1_diagnosis_voice_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_diagnose_voice_api_v1_diagnosis_voice_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceDiagnosisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/chat/conversations": {
      "get": {
        "tags": [
          "Medical Chatbot"
        ],
        "summary": "List Conversations",
        "operationId": "list_conversations_api_v1_chat_conversations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatConversationOut"
                  },
                  "type": "array",
                  "title": "Response List Conversations Api V1 Chat Conversations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/conversations/{conversation_id}/messages": {
      "get": {
        "tags": [
          "Medical Chatbot"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_api_v1_chat_conversations__conversation_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessageOut"
                  },
                  "title": "Response List Messages Api V1 Chat Conversations  Conversation Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/send": {
      "post": {
        "tags": [
          "Medical Chatbot"
        ],
        "summary": "Send Message",
        "operationId": "send_message_api_v1_chat_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_diagnose_image_api_v1_diagnosis_image_post": {
        "properties": {
          "body_part": {
            "$ref": "#/components/schemas/ImageBodyPart"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "body_part",
          "file"
        ],
        "title": "Body_diagnose_image_api_v1_diagnosis_image_post"
      },
      "Body_diagnose_voice_api_v1_diagnosis_voice_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_diagnose_voice_api_v1_diagnosis_voice_post"
      },
...
      "ChatConversationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at"
        ],
        "title": "ChatConversationOut"
      },
      "ChatMessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/ChatRole"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "retrieved_sources": {
            "items": {
              "$ref": "#/components/schemas/RetrievedSource"
            },
            "type": "array",
            "title": "Retrieved Sources"
          },
          "used_fallback": {
            "type": "boolean",
            "title": "Used Fallback"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "used_fallback",
          "created_at"
        ],
        "title": "ChatMessageOut"
      },
      "ChatRole": {
        "type": "string",
        "enum": [
          "user",
          "assistant"
        ],
        "title": "ChatRole"
      },
      "ChatSendRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatSendRequest"
      },
      "ChatSendResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "user_message": {
            "$ref": "#/components/schemas/ChatMessageOut"
          },
          "assistant_message": {
            "$ref": "#/components/schemas/ChatMessageOut"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "user_message",
          "assistant_message"
        ],
        "title": "ChatSendResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageBodyPart": {
        "type": "string",
        "enum": [
          "skin",
          "eye",
          "tongue",
          "nails",
          "throat",
          "wound",
          "other"
        ],
        "title": "ImageBodyPart"
      },
      "ImageDiagnosisOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "body_part": {
            "$ref": "#/components/schemas/ImageBodyPart"
          },
          "quality_ok": {
            "type": "boolean",
            "title": "Quality Ok"
          },
          "quality_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Note"
          },
          "predicted_disease": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Disease"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "body_part",
          "quality_ok",
          "quality_note",
          "predicted_disease",
          "confidence_score",
          "risk_level",
          "explanation",
          "created_at"
        ],
        "title": "ImageDiagnosisOut"
      },
...
      "PredictionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/PredictionSource"
          },
          "input_summary": {
            "type": "string",
            "title": "Input Summary"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/DiseasePredictionItem"
            },
            "type": "array",
            "title": "Results"
          },
          "top_disease": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Disease"
          },
          "top_probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Probability"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "recommended_specialist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Specialist"
          },
          "recovery_estimate_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Estimate Days"
          },
          "explainability": {
            "$ref": "#/components/schemas/ExplainabilityInfo"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "input_summary",
          "results",
          "top_disease",
          "top_probability",
          "risk_level",
          "recommended_specialist",
          "recovery_estimate_days",
          "explainability",
          "created_at"
        ],
        "title": "PredictionOut"
      },
      "PredictionSource": {
        "type": "string",
        "enum": [
          "text",
          "voice",
          "image",
          "lab_report"
        ],
        "title": "PredictionSource"
      },
      "ProgressionOut": {
        "properties": {
          "series_id": {
            "type": "string",
            "title": "Series Id"
          },
          "body_part": {
            "type": "string",
            "title": "Body Part"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ProgressionPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          }
        },
        "type": "object",
        "required": [
          "series_id",
          "body_part",
          "points",
          "trend"
        ],
        "title": "ProgressionOut"
      },
      "ProgressionPoint": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "predicted_disease": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Disease"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "created_at",
          "predicted_disease",
          "confidence_score",
          "risk_level"
        ],
        "title": "ProgressionPoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceDiagnosisOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "extracted_symptoms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extracted Symptoms"
          },
          "prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "transcript",
          "created_at"
        ],
        "title": "VoiceDiagnosisOut"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}